requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
fire==0.3.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
Pillow==7.2.0
proglog==0.1.9
pylint==2.6.0
requests==2.24.0
six==1.15.0
termcolor==1.1.0
toml==0.10.1
tqdm==4.50.2
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
PySimpleGUI==4.32.1
